pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "uncms"
version = "0.0.10"
authors = [
    { name="Lewis Collard", email="lewiscollard@gmail.com" },
]
readme = "README.md"
description = "A CMS toolkit for Django emphasising simplicity, speed, and familiarity."
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
]

dependencies = [
    "Pillow",
    "sorl-thumbnail",
    "beautifulsoup4",
    "bleach>=5.0.1,<6",
    "django>=4.2,<4.3",
    "django-watson",
    "django-reversion",
    "python-magic>=0.4.15",
]

[project.optional-dependencies]
dev = [
    "black==23.12.1",
    "build==0.9.0",
    "flake8==6.0.0",
    "factory-boy==3.2.1",
    "isort==5.13.2",
    "Jinja2==3.1.2",
    "psycopg2",
    "pylint==2.15.10",
    "pylint-django==2.5.3",
    "pytest==7.2.0",
    "pytest-cov",
    "pytest-django==4.5.2",
    "twine==4.0.2",
]

jinja2 = [
    "Jinja2>=3.1.2,<4",
]

testhelpers = [
    "factory-boy==3.2.1",
]

[project.urls]
"homepage" = "https://uncms.dev/"
"documentation" = "https://docs.uncms.dev/"
"repository" = "https://github.com/uncms-dev/uncms"

[tool.coverage.run]
omit = ["*/migrations/*"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = ["--create-db", "--no-migrations"]
# Turn warnings into errors
filterwarnings = [
    "error",
    # except this one from deep within Sorl :(
    "ignore:unclosed file:ResourceWarning",
]
